requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
python-dotenv
Flask-Bcrypt
Flask-Migrate
pydantic
